fix(sales & inventory): secure selling price logic, fix stock decimal precision, and add fluent validation pipeline

+++ NexusErp.Application/Procurement/Queries/GetProductStockQuery.cs
@@ -9,5 +9,5 @@ namespace NexusErp.Application.Procurement.Queries
     public record GetProductStockQuery(
         Guid BranchId,
         Guid ProductId
-        ) : IRequest<Result<int>>;
+        ) : IRequest<Result<decimal>>;
 }
